maint: reorganize tests around the spun-off apply
